Adding ServerJsonSerializer to save world/player/base data as json (#1075)
* Rename ServerProtobufSerializer to ServerProtoBufSerializer

* Added option to save world/player/base data as json

* Renaming "SubnauticaServerProtobufSerializer" to "SubnauticaServerProtoBufSerializer"

* Adding json serialization to tests

* > Added NitroxId Converter
> Moved JsonConverters one folder up

* ServerConfig changes

* Fixing serialization tests

* Making all constructor without parameter in GameLogic protected so json uses the constructor with parameter

* Making ProtoBuf standard serializer

* Making ProtoBuf standard serializer

* Moving converter to NitroxServer

* Adding option to backup world if deserialization fails.

* Fixing errors after rebase

* Switching from custom Newtonsoft.Json to nuget package

* Minor fixes

* Added option for serialize/deserialize directly with file path

* Only backup save if all files are there

+++ NitroxServer/GameLogic/Bases/BaseData.cs
@@ -7,14 +7,14 @@ namespace NitroxServer.GameLogic.Bases
     [ProtoContract]
     public class BaseData
     {
-        public const long VERSION = 2;
-        
+        public const short VERSION = 2;
+
         [ProtoMember(1)]
         public List<BasePiece> PartiallyConstructedPieces = new List<BasePiece>();
 
         [ProtoMember(2)]
         public List<BasePiece> CompletedBasePieceHistory = new List<BasePiece>();
-                
+
         public static BaseData From(List<BasePiece> partiallyConstructedPieces, List<BasePiece> completedBasePieceHistory)
         {
             BaseData baseData = new BaseData();
